+++ android/src/main/java/com/reactcommunity/rndatetimepicker/RNDatePickerDialogFragment.java
@@ -131,11 +131,11 @@ public class RNDatePickerDialogFragment extends DialogFragment {
 
     if (args != null) {
       // for dialog buttons
-      if (args.containsKey(RNConstants.ARG_POSITIVE_BUTTON_LABEL)) {
-        dialog.setButton(DatePickerDialog.BUTTON_POSITIVE, args.getString(RNConstants.ARG_POSITIVE_BUTTON_LABEL), dialog);
+      if (args.containsKey(RNConstants.ARG_CONFIRM_TEXT_ANDROID)) {
+        dialog.setButton(DatePickerDialog.BUTTON_POSITIVE, args.getString(RNConstants.ARG_CONFIRM_TEXT_ANDROID), dialog);
       }
-      if (args.containsKey(RNConstants.ARG_NEGATIVE_BUTTON_LABEL)) {
-        dialog.setButton(DatePickerDialog.BUTTON_NEGATIVE, args.getString(RNConstants.ARG_NEGATIVE_BUTTON_LABEL), dialog);
+      if (args.containsKey(RNConstants.ARG_CANCEL_TEXT_ANDROID)) {
+        dialog.setButton(DatePickerDialog.BUTTON_NEGATIVE, args.getString(RNConstants.ARG_CANCEL_TEXT_ANDROID), dialog);
       }
     }