+++ src/index.d.ts
@@ -122,8 +122,8 @@ export type AndroidNativeProps = Readonly<
 
       onChange?: (event: AndroidEvent, date?: Date) => void;
       neutralButtonLabel?: string;
-      positiveButtonLabel?: string;
-      negativeButtonLabel?: string;
+      confirmTextAndroid?: string;
+      cancelTextAndroid?: string;
     }
 >;