+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 2.7.0
 
 - Add support for 'locale' prop for Android.
-- Add 'positiveButtonLabel' and 'negativeButtonLabel' props to pass Android's DateTimePicker dialog buttons.
+- Add 'confirmTextAndroid' and 'cancelTextAndroid' props to pass Android's DateTimePicker dialog buttons.
 
 ### 2.6.0